rename: RushGame 관련 폴더명 변경
