+++ client/src/pages/RushGame/index.tsx
@@ -1,10 +1,10 @@
 import { motion } from "framer-motion";
 import CTAButton from "@/components/CTAButton";
 import { ASCEND, SCROLL_MOTION } from "@/constants/animation.ts";
-import CardOptions from "@/features/RushGame/RushGameSection/CardOptions.tsx";
-import Countdown from "@/features/RushGame/RushGameSection/Countdown.tsx";
-import FinalResult from "@/features/RushGame/RushGameSection/FinalResult.tsx";
-import SelectedCard from "@/features/RushGame/RushGameSection/SelectedCard.tsx";
+import CardOptions from "@/features/RushGame/RushGameSections/CardOptions.tsx";
+import Countdown from "@/features/RushGame/RushGameSections/Countdown.tsx";
+import FinalResult from "@/features/RushGame/RushGameSections/FinalResult.tsx";
+import SelectedCard from "@/features/RushGame/RushGameSections/SelectedCard.tsx";
 import { useRushGameContext } from "@/hooks/useRushGameContext.ts";
 
 // TODO: 계속 카운트 다운에 맞춰 매초 렌더링 되는 문제 해결
